requirements.txt
Flask==3.0.3
keras==3.5.0
tensorflow-cpu==2.15.0
nltk==3.9.1
numpy==1.23.5
matplotlib==3.8.4
gunicorn==21.2.0